Add dependencies and move tests for discoverability
