+++ bgbb/bgbb_likelihood.py
@@ -116,14 +116,3 @@ def nb_loglikelihood_df(params, df, rfn_names=False, para=True, nll=False, ncust
         return -np.mean(ll * ncusts)
     return ll
 
-
-#########
-# Tests #
-#########
-def test_nb_loglikelihood():
-    res = nb_loglikelihood([0.1, 0.1, 0.2, 0.3], np.r_[3], np.r_[4], np.r_[5])
-    assert np.allclose(res, [-6.31032992])
-
-
-# test_nb_lbeta()
-# test_nb_lbeta_vec2()
